correct overflow for > 10^9 number

+++ algorithmic-toolbox/Assignment-2_Algorithmic_Warm-up/2.1.4_LastDigitOfTheSumOfFibonacciNumbers/FinalVersionLastDigitOfTheSumOfFibonacci.cpp
@@ -56,7 +56,7 @@ int FastLastFibNumber(long long n)
 
 int main()
 {
-    int Fib_n;
+    long long Fib_n;
     std::cin >> Fib_n;
 
     int result = FastLastFibNumber(Fib_n);
